Spits branch and status nodes into separate files Removes status node for now
