+++ src/views/explorerNodes.ts
@@ -1,9 +1,12 @@
 'use strict';
 
 export * from './explorerNode';
+export * from './branchesNode';
+export * from './branchHistoryNode';
 export * from './commitFileNode';
 export * from './commitNode';
 export * from './fileHistoryNode';
 export * from './repositoryNode';
 export * from './stashCommitNode';
 export * from './stashNode';
+export * from './statusNode';
